requirements.txt
Appium-Python-Client==0.48
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
idna==2.8
importlib-metadata==0.23
more-itertools==7.2.0
packaging==19.2
pathlib2==2.3.5
pkg-resources==0.0.0
pluggy==0.13.1
py==1.8.0
pyparsing==2.4.5
pytest==5.3.0
pytest-base-url==1.4.1
pytest-html==1.22.1
pytest-metadata==1.8.0
pytest-selenium==1.17.0
pytest-variables==1.9.0
requests==2.22.0
six==1.13.0
urllib3==1.25.7
wcwidth==0.1.7
zipp==0.6.0
